pyproject.toml
[tool.poetry]
name = "lynxie"
version = "0.1.0"
description = ""
authors = ["Fluffy <michal-gdula@protonmail.com>"]
readme = ".github/README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
black = "^23.7.0"
sqlalchemy = "^2.0.20"
python-dotenv = "^1.0.0"
discord-py = {extras = ["voice"], version = "^2.3.2"}
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
selenium = "^4.12.0"
webdriver-manager = "^4.0.0"
flask = "^2.3.3"
yt-dlp = "^2023.7.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
